Explicitly parse 'independent' config as boolean (#7)
This commit fixes an issue where the 'independent' config is
parsed and evaluated as a string. This causes counter-intuitive
behaviour when supplying a falsy value to the configuration, such as

    independent = 0

This would previously be parsed as a string, and since '0' is a truthy
value, the independent behaviour would be enabled.

The issue is fixed by parsing the configuration explicitly as a boolean.